"LA.UM.5.5.r1-00900-8x96.0"

+++ src/com/android/bluetooth/opp/BluetoothOppObexServerSession.java
@@ -101,7 +101,7 @@ public class BluetoothOppObexServerSession extends ServerRequestHandler implemen
 
     boolean mTransferInProgress = false;
 
-    private int position;
+    private long position;
 
     public BluetoothOppObexServerSession(Context context, ObexTransport transport) {
         mContext = context;
